requirements.txt
django
django-unfold
pillow
django-ckeditor
